feat : retrive data from api and login cache in profile, split setting and profile page
