+++ lib/component/data/GlobalVariable.dart
@@ -1,4 +1,4 @@
-const String BASE_URL = "https://539c-112-78-156-225.ngrok-free.app/api";
+const String BASE_URL = "https://b57e-112-78-156-227.ngrok-free.app/api";
 String? token = "";
 
 Map<String, String> loginData = {
